product buydetail css 적용전

+++ src/main/java/com/greenright/dao/ReviewDao.java
@@ -8,5 +8,5 @@ public interface ReviewDao {
   int checkReview (Review review) throws Exception;
   List<Review> findAll(int ProductNo)throws Exception;
   Review findBy(int no) throws Exception;
-  int getRatingAver(int no) throws Exception;
+  double getRatingAver(int no) throws Exception;
 }
